Replace `Display` impls with `display_more`
This commit uses the `display_more` crate to replace existing
`Display` impls with procedural macros to generate the impls.
This reduces the size of the source code with a minimal increase
in compilation times. The benefit here is really to make the
source code easier to read, as most of the display impls are
simple or even just delegating to the internal type which they
wrap, as we use the newtype pattern pretty aggressively.