Change output path logic to only write to file.
Realized that writing to `stdout` by default doesn't work because
the SPDX file includes a field which specifies its own file name.
This commit reworks the logic to instead choose a default file
name if none was provided by the user, and in either case to
ensure the `DocumentName` field is set correctly in the SBOM.

This also adds the `-F`/`--force` command to overwrite any
existing file at the output path.