Error out earlier for unimplemented formats.
Previously, `cargo spdx` would create the output file prior to
checking if the output format was supported. This would result in
an empty file being left over for formats which were not supported,
which isn't ideal.

This change moves the format support check to the CLI argument
parsing stage, ensuring errors arise early.