Added new 'format' module.
This commit adds a new module called `format` which includes the
definition of the `Format` type and will be the parent module
for all handling of output to the different supported output
formats. Right now that only includes the 'kv' / 'Key-Value'
format, which has been moved into this new module.