Assorted refactoring.
This commit renames some structs, moves some code between modules,
and generally tries to do a couple things. First, make the `run`
function clearer to read; it loads the info we need, builds up
the SPDX document, and then writes it out to disk. Second, to
better "tuck away" the more fiddly bits like handling conditions
for output or figuring out the output file extension. The details
of how these work involves a bit of visual complexity that most
of the time would be irrelevant to understanding the code where
they're used, so they've been refactored to reduce that cognitive
load for future readers of the code.