Vendor libgit2 in, rather than using from system.
To make this thing easier to build on people's own platforms,
and to enable more consistent behavior, let's just vendor stuff
in. Eventually it would be great to vendor everything, though
for now it looks like libgit2 picks up zlib from the user's
system, and I'm not certain how to avoid that.