Reduced TOCTOU issue with output file.
This commit moves checks related to the output file to now occur
immediately prior to writing to the file. This helps avoid issues
where we might check the output path at the beginning of execution,
during parsing, and it passes, only for the file to be deleted and
replaced with a directory before we reach writing.

This type of error is still _possible_, but this change reduces
the window of time for this error to arise.