Added basic logging.
This commit adds `log` and `env_logger` and depenedencies, and
adds some simple log statements at the 'info' level to say what
is happening as the command runs. In the future this can be used
to provide more detailed information about what inferences
'cargo-spdx' is making as it runs, or to provide warnings where
it is having to proceed with partial information.