Added interactivity, with optional non-interactive mode.
This commit adds the ability for fields to be prompted for in an
interactive fashion, which should make the command easier to use.
There's also now a --no-interact/-n flag, which turns off
interactive mode. If interaction is turned off, and required
fields are not specified, then 'cargo-spdx' will error out with
a message explaining the issue.